services/tts-adapter/pyproject.toml
[project]
name = "tts-adapter"
version = "0.1.0"
description = "OpenAI-compatible local TTS adapter service"
readme = "README.md"
authors = [
  { name = "Codex Agent", email = "codex-agent@example.com" }
]
requires-python = "==3.11.*"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "numpy>=2.3.0",
  "huggingface_hub>=0.34.0",
  "safetensors==0.5.3",
  "chatterbox-tts>=0.1.2",
  "torch==2.6.0",
  "torchaudio==2.6.0"
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "pytest-asyncio>=1.1.0",
  "httpx>=0.28.0",
  "mypy>=1.17.0",
  "ruff>=0.12.0"
]

[project.scripts]
tts-adapter-api = "tts_adapter.main:run_server"
tts-adapter-cli = "tts_adapter.cli:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N", "ANN", "RUF"]
ignore = []

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true
pretty = true
plugins = ["pydantic.mypy"]
follow_imports = "skip"
disallow_subclassing_any = false
disallow_any_decorated = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["chatterbox.*", "safetensors.*", "huggingface_hub.*", "torchaudio.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.extra-build-dependencies]
pkuseg = ["numpy>=2.3.0"]
